Rename module to match scss component
Renaming `menu` to `header`
